refactor: composition 레이어 제거
